Shift unsigned where the value reaches the sign bit
A sweep of all 599 modes under UndefinedBehaviorSanitizer reports 51 times from
11 distinct places. 38 of the 51 are one line, and the rest are the same shape
written out in eight parsers.

The one line is in the kernel helper every hash function reaches through
append_0x80_*, so it fires for any mode whose parser hashes anything:

    const u32 r = 0xff << ((offset & 3) * 8);

0xff is an int, so the last of the four offsets shifts it into the sign bit.
0xffu is the same number with the type the expression already assumes.

The others are a u8 shifted past bit 24, and one 1 << 31:

    oldoffice01->rc4key[0] |= hex_convert (rc4key_pos[0]) << 28;   m09720, m09820, m10420
    digest[1] = (t[7] << 29) | (t[8] << 24) | ...                  m28503, m28504, m30903, m30904
    salt->scrypt_N = 1 << N_log2;                                  m36300

The base32 unpackers take the cast on every term rather than the three lines
the sanitizer happened to reach, so the shape is gone from the file rather than
from one expression in it.

    CFLAGS="-g -O1 -fsanitize=fuzzer-no-link,address,undefined -fno-omit-frame-pointer" \
      tools/fuzz/build.sh
    UBSAN_OPTIONS=halt_on_error=1 ./fuzz_out/fuzz_parse_2811 fuzz_out/work/seeds/parse_2811

    before: OpenCL/inc_common.cl:4125:22: runtime error: left shift of 255 by 24
            places cannot be represented in type 'int'
    after:  no report

Every mode touched still cracks its own example hash. Two findings from the
same sweep are left for their own commits, because they are bugs rather than
promotions: a shift exponent taken from the line in m14200, and an index past
the end of a u32[64] in m29100.

+++ src/modules/module_09820.c
@@ -190,7 +190,7 @@ int module_hash_decode (MAYBE_UNUSED const hashconfig_t *hashconfig, MAYBE_UNUSE
   oldoffice34->rc4key[1] = 0;
   oldoffice34->rc4key[0] = 0;
 
-  oldoffice34->rc4key[0] |= hex_convert (rc4key_pos[0]) << 28;
+  oldoffice34->rc4key[0] |= (u32) hex_convert (rc4key_pos[0]) << 28;
   oldoffice34->rc4key[0] |= hex_convert (rc4key_pos[1]) << 24;
   oldoffice34->rc4key[0] |= hex_convert (rc4key_pos[2]) << 20;
   oldoffice34->rc4key[0] |= hex_convert (rc4key_pos[3]) << 16;
